requirements.txt
Flask>=2.0
python-dotenv>=1.0
web3>=6.0
eth-account>=0.9
requests>=2.28
discord.py>=2.3
ens>=0.0.0  # optional, some web3/ens features depend on this
gunicorn>=20.1
